[PAGOPA-859] fix: resolved misconceptions on functions
